Update 3 recover
Added more accurate memory buffer to hole blocks read in from input file.  Changed read to buffer into a variable so it can be used later in a condition.  Added print statements for debugging.  Added condition that checks if read method is returning succesful read memory block.  If not end of file is found and image file is closed and loop ended.  Added write next 512 bytes to file if not end of file.  Working as expected.  The program now reads the raw memory card file,  Breaks it down into chunks of 512 bytes and looks for the intro headers for a jpeg.  If it finds the intro headers it copies the 512 bytes into a new blank jpeg file in working dir.  If it doesn't find the intro header and a jpeg hasn't been found it discards the bytes and continues searching for a jpeg.  The program correctly now recovers all 50 jpegs!