Update 3 speller
Added logic for adding word from dictionary into hashtable.  Added logic for adding first node into hash table to start the linked list.  Began work on modifying pointers to create the chain of nodes when a new node is added.  The first node will point to the new node and the new node will point to the node the first node was pointing to.